Removed need for link_id on assessment_scores table, some bugs
Resource_link_id isn't necissary on the assessment_scores table.
It has a 1:N relationship with attempts that has that field, so I
removed the migration to add it and updated the necissary queries.

Also found a bunch of inconsiquential cases where isPreview was not taken
into account properly in different scoring queries.  This appeared to only
affect events in preview mode, but there's no need to hard code isPreview, so
I updated thos methods to accept it as an arg.