Refactoring by splitting up and moving
Server side rendering of the repository is a little complex.
Before this chagne we used a file sitting in /client/name.js
to represent a webpack entry point.

This reduces complexity by using the built in mechanism to combine scripts
into a webpack output file serverScripts config option. No need for extra files.

* removes repository/server/dashboard.js (and more) - rely on serverScripts config instead
* Moves all SSR 'Page' components into a page subdirectory to isolate them a bit.
* Renames layout and page components so that class names mirror the file name better.
* Updates readme in repository and module-selector packages
* splits apart assessment/server/express into events, services, and more isolated express parts
* removes completely unused repository components (they were part of early prototypes)