[CLEAN] Change getCurrentUser & requireCurrentUser
* getCurrentUser no longer has an option to throw if user is missing
* use requireCurrentUser instead of getCurrentUser(true)

req.getCurrentUser used to support an option that would cause it to
an error if the current user is not a logged in (non-guest) user.
A precise name for that option would be extra verbose:
ex: throwErrorIfNotLoggedInOrIsGuest

As a solution, I looked at changing getCurrentUser to
getCurrentUserOrGuest, but we want to make sure getCurrentUser looks
like the default option for this behavoir. Keeping it simple seems
important.

I realized the option on getCurrentUser was ONLY used for
requireCurrentUser. So if we move the implementation over to that
method, it moves implementation closer to it's purpose.