[fix] math equation editor no longer jumps to end
* Math Equation editor inputs held in state
* Math Equation now debounces updates to the model for speed
* Math Equation tests upgraded a little bit
* default values added for math equation

The math equation editor had an issue where typing into the editors
would always jump the cursor to the end of the input. This is generally
caused by controlled inputs not updating fast enough.  In this case, we
were updating the model and waiting for the display to re-render.  I
changed the component to render and update state with a debounced update
to the model.