[fix] optimize qb editor settings component
The number input for chooseAmount was updating the slate model
on changes.  Text input boxes aren't too friendly when controlled by
React slowly - the cursor jumps to the end of the input - so this changes
the values to be localized in state and periodically broadcast out to the
slate model.

* moved calls to .bind out of render
* inputs are handled in state locally and passed out to model via debaunced method
* tests updated where appriopriate