[fix] unlocked content breaks assessment question context
When the content section isn't locked durring an assessment, if the
student navigated to a content page durring assessment, NavUtil would
be told about the transition out of assessment contenxt, but not told
about the transition back into assessment context.

This fix has the assessment viewer react component announce the context
swap when it's mounted on the dom.