cleanup of rename and copy options in editor
The update/save title options were very complicated and needed some updates to simplify.
They do work a little differently in code-editor vs page-editor, but they seem to make sense now.
Copying a module used to be implemented entirely on the client side, this will now use
the route from obojobo-respository to copy a module, simplifying the client and allowing
use to use the server's metadata to track copies better.

* editor-title-input keeps temporary state internally now
* code-editor save title fixes
* file-menu removes rename option
* file-menu 'copy module' now uses obojobo-repository copy route instead of handling copy on the client
* /api/drafts/xxx/copy now accepts a title param and defaults to appending copy to the name
* page-editor update title fixes