[fix] survey question editor toggle bug fixed
In the editor, toggling the survey only button would cause an
error when the question had a Solution child. The slate transform
didn't take into account that the question could have a solution.

The fix searches for the MCAssessment node instead of assuming it's
the last child of the question

fixes #1319