[fix] visual editor incorrectly restores module
per issue #1330 sometimes the visual editor could
restore an old version of the module unintentionally.
This would happen if a drastic change introduced in the
code editor, then switched back to visual, and you edit
the module's title in the header.

The problem stems from an issue I discovered some time ago
with OboModel not being friendly if you try to load a 2nd
draft module.  It breaks because of OboModel.getRoot was
not built to be able to differentiate between multiple root
models.

This fix adds an option to reset OboModel's state so you can
load a new model.

For this issue, that comes into play when switching from xml
or json back to the visual editor & renaming. This process will
no longer return references to old nodes.