mathEquation editor saves more often
MathEquation would previously only save when the edit dialog
switched from open to closed.  Unfortunately it doesn't close
when you click anywhere that's not in the slate display.

Ex: edit contents of the equation and click save or preview from the
toolbar.  The contents aren't copied to the slate document and when
save is called, the changes don't make it into the document

This adds a debounced function to make sure the document is saved
as the content changes.

It also second condition in componentDidUpdate that'll cause the
contents to be copied to slate - when the editor is closing (previously
was just when it lost selection)

At the same time I moved the code that closes the editor out
of the function intended to push updates to slate.