fixup rebase of questionbank fixes onto dev11
This is an attempt to consolidate an older branch
with changes that happened in question bank.

QuestionBank's editor component was converted into a
functional component which changed the original intent
of the pull request - HOWEVER the pull request contained
a good number of other fixes and optimizations.  So this is
my attempt con consolidate those.