simplify ActionButtonEditorAction key
We needed a more uniqe key, which the previous commit provided,
but it is now unnecessarily complex.  We'll switch to just using
their array index as the key.

I believe they were previously a combination key to prevent them
from overlapping with other lists, but react keeps keys local
to it's siblings - so that shouldn't be a problem here.