[fix] moreinfobox causing errors on slate undo/redo
MoreInfoBox was causing slate/react rendor errors when
using ctrl+z keys if the slate editor wasn't focused.

The problem stemed from morinfobox calling it's onBlur prop
when being removed from the dom (.close function) which would
then cause other problems.  This commit prevents close from calling
the onBlur prop if it's not actually open.

BONUS 1: Simplified moreinfobox's mousedown listener - it was
being added and removed enough to make it difficult to follow / easy
to lose track of.  Now the handler uses the components state to determine
what to do, and it's much easier to keep track of.

BONUS 2: The visual editor error boundry component tries to call
undo when there's a render error.  In the case of the problems fixed
by this pull request - the undo call being used to recover from the error
would throw an error.  This wasn't caught so, I added that hoping
that it'll keep the editor from white screen-ing.

In the future - we need to capture those and send them to a logging service