adds user_id to draft_content *migration*
Draft_content didn't keep track of who created it. This is mostly
due to the fact that it previously wasn't possible to have multiple
users edit a module.  Now that sharing was added, we're lacking
the ability to know who saved which version of a draft.  Which is
especially needed in this PR - reverting / restoring old versions.

* adds a migration to add user_id to drafts_content
* updates queries to insert the correct user_id
* cleans up a migration with mixed tab indenting
* fixes revert / restore query that would return random owner