[fix] obojobo-migrate compatable w config default
The db.json file changed to conform to the rest of our
config file formats, allowing a default option that gets
added to for the given environment.  The problem, is
db-migrate uses this file directly and doesn't work the same
way - if default is set, it just uses it.

So this pr lets obojobo-migrate set the default to development,
but still allows -e and -env flag to set the environment.

New bonus, a console log about the environment being used was added
for help when someone has issues.