much more reliable close window dispatcher
sendBeacon wasn't reliably deleting an edit lock when closing
the window.  The problem is that onBeforeUnload is too late to
rely on any requests to be queued before closing.  The answer
is a combo of the visibility api and iOS specific events, with
a fallback to onbeforeunload.

see https://volument.com/blog/sendbeacon-is-broken for some
description of the issue/problem (the comments are key)

This new code still uses beforeunload, but adds pagehide and
visibilitychange if it's available.