[fix] module selector no-items not being cleared
In the module selector, when the user had no
modules of their own, opened my modules, then opened
the community library - the warning about not having any
modules would remain open. This is due to the query that
was hiding it not taking into account that there were
multiple matching dom elements.

Fix is to hide them all before unhiding.