adds viewer:initialView event and fixes opening in a new tab
Problem: if you use the dev dashboard to launch a student view
in a new tab, it'll launch that tab without switching focus to it.

Once you switch to that tab, obo thinks you must have left and
come back, so the code tries to send a return event - with a
reference to a non-existent leave event.

This commit adds a new event to represent the first time the
module is visible to the student.  It will increase timing
accuracy and prevent the error thats currently being thrown
due to an undefined leave event reference.