adds insertItemInto to Figure node plugins and..
To modularize code for inserting using the file toolbar
the nodes need to be able to register to handle insert events

this adds insertItemInto to Figure and changes file-toobar-viewer
to support the changes.

Further more, to modularize file-toolbar-viewer, each node
can now register if it accepts insert requests `acceptsInserts`.

Question, QuestionBank, and Table have been added to the list that
do not support inserts.  This moves the specific implementation for
those out of the toolbar and into the chunks themselves.